refactor: create and use routes enum
