+++ weather-app/src/components/Navbar.tsx
@@ -1,11 +1,12 @@
 import { NavLink } from "react-router";
+import { Routes } from "../routes";
 
 export const Navbar = () => {
   return (
     <nav className=" bg-amber-100 flex justify-around py-3.5 font-bold">
-      <NavLink to={"/"}>Home</NavLink>
-      <NavLink to={"/forecasts"}>All forecasts</NavLink>
-      <NavLink to={"/favorites"}>Favorites</NavLink>
+      <NavLink to={Routes.HOME}>Home</NavLink>
+      <NavLink to={Routes.FORECASTS}>All forecasts</NavLink>
+      <NavLink to={Routes.FAVORITES}>Favorites</NavLink>
     </nav>
   );
 };
