+++ weather-app/src/routes.ts
@@ -1,0 +1,5 @@
+export const enum Routes {
+  "HOME" = "/home",
+  "FORECASTS" = "/forecasts",
+  "FAVORITES" = "/favorites",
+}
